+++ pkg/p2p/node.go
